% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.auxiliary.functions.R
\name{flag.nameiftrue}
\alias{flag.nameiftrue}
\title{flag.nameiftrue
Aux function for parsing filenames. Returns a variable's name if its value is TRUE.}
\usage{
flag.nameiftrue(toggle, prefix = NULL, suffix = NULL, name.if.not = "")
}
\arguments{
\item{toggle}{Variable name to return as string, if its value is TRUE.}

\item{prefix}{prefix}

\item{suffix}{suffix}

\item{name.if.not}{Alternative name to return, "" by default.}
}
\value{

}
\description{
flag.nameiftrue
Aux function for parsing filenames. Returns a variable's name if its value is TRUE.
}
\examples{
a=TRUE; b=FALSE; flag.nameiftrue(a); flag.nameiftrue(b)
}
