% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_transform.R
\name{fourier_transform}
\alias{fourier_transform}
\title{Creates a Fourier transform plot of the data specified in the 'data_col' column using ggplot2}
\usage{
fourier_transform(data, time_col, data_col)
}
\arguments{
\item{data}{DataFrame containing all necessary information}

\item{time_col}{string containing the time data column corresponding to the data we
want to apply Fourier transform to}

\item{data_col}{string containing the name of the column we want to apply Fourier transform to}
}
\value{
a ggplot2 line plot of frequency vs amplitude of the data inputed
}
\description{
Creates a Fourier transform plot of the data specified in the 'data_col' column using ggplot2
}
\examples{
fourier_transform(data = my_data,
                  time_col = 'time_series',
                  data_col = 'signal')

}
