% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Make a ggplot histogram}
\usage{
gghist(data, variable)
}
\arguments{
\item{data}{A tibble or dataframe.}

\item{variable}{The column name inside data that will be made into a histogram.}
}
\value{
Produces a ggplot histogram with mean, median and standard deviation indicated on the plot.
}
\description{
Creates a ggplot histogram that indicates the position of the mean and median and displays the standard deviation.
}
\examples{
gghist(iris, Sepal.Length)
}
