% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.auxiliary.functions.R
\name{kpp}
\alias{kpp}
\title{Collapse vector to a string, separated by dots}
\usage{
kpp(...)
}
\arguments{
\item{...}{Pass any other parameter of the corresponding plotting function(most of them should  work).}
}
\description{
Collapse vector to a string, separated by dots
}
\examples{
kpp(1,3, 'sdasd'); kpp(1:3, 'sdasd')
}
