% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{q32vA4_grid_plot}
\alias{q32vA4_grid_plot}
\title{q32vA4_grid_plot}
\usage{
q32vA4_grid_plot(
  plot_list,
  suffix = NULL,
  plotname = FixPlotName(substitute(plot_list), suffix),
  plot = FALSE,
  nrow = 3,
  ncol = 2,
  extension = c("pdf", "png")[2],
  scale = 1,
  h = get0("hA4", ifnotfound = 11.69) * scale,
  w = get0("wA4", ifnotfound = 8.27) * scale,
  ...
)
}
\arguments{
\item{plot_list}{A list of ggplot objects, each of which is one panel.}

\item{suffix}{A suffix added to the filename. Default: NULL}

\item{plotname}{Plot name. Default: Autonaming.}

\item{plot}{Show the plot? Default: F}

\item{nrow}{number of rows for panels on the page. Default: 2}

\item{ncol}{number of columns for panels on the page. Default: 2}

\item{extension}{file extension}

\item{scale}{Scaling factor of the canvas. Default: 1}

\item{h}{Height of the plot, in inches. Default: get0("hA4", ifnotfound = 11.69) * scale}

\item{w}{Width of the plot, in inches. Default: get0("wA4", ifnotfound = 8.27) * scale}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot up to 6 panels (3-by-2) on vertically standing A4 page.
}
\examples{
# q32vA4_grid_plot()
}
