% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot.stacked.from.wide.df}
\alias{qbarplot.stacked.from.wide.df}
\title{qbarplot.stacked.from.wide.df - Barplot for tibbles or dataframes}
\usage{
qbarplot.stacked.from.wide.df(
  df,
  x = "Samples",
  y = "Fraction",
  z = "Category",
  y_axis_factor_levels = colnames(df),
  plotname = FixPlotName(substitute(df)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  plot = TRUE,
  save = TRUE,
  ext = "png",
  also.pdf = FALSE,
  save.obj = getOption("gg.save.obj", FALSE),
  color = 1,
  label = NULL,
  scale = TRUE,
  hline = FALSE,
  palette_use = getOption("gg.palette_use", "jco"),
  xlab.angle = 45,
  xlab = x,
  logY = FALSE,
  annotation_logticks_Y = logY,
  legend.position = "none",
  max.names = 50,
  limitsize = FALSE,
  max.categ = 10,
  grid = "y",
  mdlink = getOption("gg.mdlink", FALSE),
  w = qqqAxisLength(df, factor = 0.7),
  h = 5,
  ...
)
}
\arguments{
\item{df}{The variable to plot.}

\item{x}{Colname to split along X axis. Default: "Samples".}

\item{y}{Colname to count along y axis. Default: "Fraction".}

\item{z}{Colname to split along Y axis. Default: "Category.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{color}{Color (split) by along Y.}

\item{label}{Label text.}

\item{scale}{Scale the y axis. Default: TRUE.
\code{MarkdownReports::setup_MarkdownReports()}, located at \code{path_of_report}. Default: FALSE}

\item{hline}{Draw a horizontal line on the plot.}

\item{palette_use}{Color palette to use. Either a \code{ggpubr::get_palette} palette or a custom vector of colors. Default: 'jco'.}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{xlab}{X-axis label. Default: \code{x}.}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{legend.position}{Character indicating the position of the legend. Default is is 'none' to hide the legend.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{Limit size.}

\item{max.categ}{The maximum allowed number of unique categories.

@param mdlink Insert .pdf and .png image links to a markdown report file, setup by
\code{MarkdownReports::setup_MarkdownReports()}, located at \code{path_of_report}. Default: FALSE}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Draw and save a stacked barplot for each row of a dataframe.
}
\examples{
# Example of wide-format data for stacked bar plot
df.SingletSplit <- data.frame(
  doublet = c(0.027886224, 0.007699141, 0.003704390, 0.003205128),
  singlet = c(0.9686280, 0.9872668, 0.9925912, 0.9119822),
  unassigned = c(0.0034857780, 0.0050340539, 0.0037043897, 0.0848126233),
  row.names = c("sc06.692", "sc06.693", "sc08.325", "sc08.327")
)

# Create the stacked bar plot using the new qbarplot.df2 function
qbarplot.stacked.from.wide.df(df.SingletSplit)

}
