% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.auxiliary.functions.R
\name{qMarkdownImageLink}
\alias{qMarkdownImageLink}
\title{Insert Markdown image link to .md report}
\usage{
qMarkdownImageLink(file_name = "myplot.pdf")
}
\arguments{
\item{file_name}{file_name}
}
\description{
Insert Markdown image link to .md report
}
\examples{
qMarkdownImageLink(file_name = "myplot.pdf")
}
