% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.auxiliary.functions.R
\name{qqqCovert.named.vec2tbl}
\alias{qqqCovert.named.vec2tbl}
\title{qqqCovert.named.vec2tbl}
\usage{
qqqCovert.named.vec2tbl(
  namedVec = 1:14,
  verbose = F,
  strip.too.many.names = TRUE,
  thr = 50
)
}
\arguments{
\item{namedVec}{namedVec}

\item{verbose}{verbose}

\item{strip.too.many.names}{strip.too.many.names}

\item{thr}{thr}
}
\description{
Covert a named vector to a table.
}
\examples{
qqqCovert.named.vec2tbl(namedVec = c("A"=2, "B"=29) )
}
