% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqList.2.DF.ggplot}
\alias{qqqList.2.DF.ggplot}
\title{qqqList.2.DF.ggplot}
\usage{
qqqList.2.DF.ggplot(ls = LetterSets)
}
\arguments{
\item{ls}{A list with all elements named}
}
\description{
Convert a list to a tow-column data frame to plot boxplots and violin plots
}
\examples{
LetterSets <- list("One" = LETTERS[1:7], "Two" = LETTERS[3:12]); qqqList.2.DF.ggplot(LetterSets)
}
