% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqTbl.2.Vec}
\alias{qqqTbl.2.Vec}
\title{qqqTbl.2.Vec}
\usage{
qqqTbl.2.Vec(tibble.input, name.column = 1, value.column = 2)
}
\arguments{
\item{tibble.input}{tibble.input}

\item{name.column}{name.column}

\item{value.column}{value.column}
}
\description{
Convert a table to a named vector.
}
\examples{
a <- 1:5
x <- tibble::tibble(a, a * 2)
qqqTbl.2.Vec(x)
}
