% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qvenn}
\alias{qvenn}
\title{Quickly draw and save a Venn Diagram (png, pdf, ggobj.qs)}
\usage{
qvenn(
  list,
  plotname = FixPlotName(substitute(list)),
  suffix = NULL,
  subtitle = paste(length(unique(unlist(list))), "elements in total"),
  caption = parseParamStringWNames(sapply(list, length)),
  caption2 = NULL,
  filename = NULL,
  plot = TRUE,
  save = TRUE,
  ext = "png",
  also.pdf = FALSE,
  save.obj = getOption("gg.save.obj", FALSE),
  col.min = "white",
  col.max = "red",
  legend.position = "none",
  x_exp = 0.2,
  mdlink = getOption("gg.mdlink", FALSE),
  w = 8,
  h = 0.75 * w,
  ...
)
}
\arguments{
\item{list}{The variable to plot.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{subtitle}{The subtitle of the plot. Default: paste (length(unique(unlist(list))), 'elements in total')}

\item{caption}{Optional text added to bottom right corner of the plot. Default is list element lengths parsed.}

\item{caption2}{Optional text added to bottom right corner of the plot. Default is NULL.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{col.min}{Color scale minimum. Default: white.}

\item{col.max}{Color scale maximum. Default: red.}

\item{legend.position}{Character indicating the position of the legend. Default is is 'none' to hide the legend.}

\item{x_exp}{Expand axis to show long set labels. Default: 0.2.}

\item{mdlink}{Insert .pdf and .png image links to a markdown report file, setup by
\code{MarkdownReports::setup_MarkdownReports()}, located at \code{path_of_report}. Default: FALSE}

\item{w}{Width of the plot.}

\item{h}{Height of the plot. Default: h = w * 0.75.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
This all-in-one function draws, annotates, displays and saves a Venn Diagram from
a named list of values. It is a wrapper around \code{ggVennDiagram::ggVennDiagram()},
with the automation of many features. All \code{ggVennDiagram} parameters can be
accessed through the \code{...} argument. It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
LetterSets <- list("One" = LETTERS[1:7], "Two" = LETTERS[3:12])
qvenn(LetterSets)

}
