% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatter.R
\name{scatter_express}
\alias{scatter_express}
\title{Create a scatterplot and calculate correlation values for two numerical variables}
\usage{
scatter_express()
}
\value{
ggplot Object
}
\description{
Creates a ggplot scatterplot object containing two numerical variables. Arguements in the
function will control whether correlational values and log transformations are calculated for the input data.
}
\examples{
scatter_express(heart_df, "chol","age", transform = FALSE, corr = "Pearson")
}
