% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Time series decomposition and plot}
\usage{
ts_plot(data, col, frequency)
}
\arguments{
\item{data}{str: the path of the csv file}

\item{col}{str: the column to be analyzed}

\item{frequency}{int: should choose from {1, 4, 12, 52}}
}
\value{
ggplot Object
}
\description{
Convert csv file into a time series object, decompose it into trend, seasonality/cyclicity
and noise/remainder/error and plot the raw data and the decomposed components.
If the time series can't be decomposed, the function will just return the line chart of the raw data.
}
\examples{
ts_plot("~/Desktop/ts4_jjshares.csv", "earnings", 4)
}
