% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{q32vA4_grid_plot}
\alias{q32vA4_grid_plot}
\title{q32_A4_plot
Plot up to 6 panels (3-by-2) on vertically standing A4 page.}
\usage{
q32vA4_grid_plot(
  plot_list,
  pname = F,
  suffix = NULL,
  plot = F,
  nrow = 3,
  ncol = 2,
  extension = c("pdf", "png")[2],
  h = hA4 * scale,
  w = wA4 * scale,
  scale = 1,
  ...
)
}
\arguments{
\item{plot_list}{A list of ggplot objects, each of which is one panel.}

\item{pname}{Plotname, Default: F}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{nrow}{number of rows for panelson the page, Default: 2}

\item{ncol}{number of columns for panelson the page, Default: 2}

\item{extension}{file extension}

\item{h}{height of the plot, Default: wA4 * scale}

\item{w}{width of the plot, Default: hA4 * scale}

\item{scale}{Scaling factor of the canvas, Default: 1}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
q32_A4_plot
Plot up to 6 panels (3-by-2) on vertically standing A4 page.
}
\examples{
# q32_A4_plot()
}
