% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qA4_grid_plot}
\alias{qA4_grid_plot}
\title{qA4_grid_plot}
\usage{
qA4_grid_plot(
  plot_list,
  plotname = FixPlotName(substitute(plot_list), nrow, "by", ncol, suffix),
  suffix = NULL,
  nrow = 3,
  ncol = 2,
  plot = FALSE,
  labels = LETTERS[1:length(plot_list)],
  max.list.length = 16,
  extension = c("pdf", "png")[2],
  scale = 1,
  h = hA4 * scale,
  w = wA4 * scale,
  ...
)
}
\arguments{
\item{plot_list}{A list of ggplot objects, each of which is one panel.}

\item{plotname}{Plot name. Default: Autonaming}

\item{suffix}{A suffix added to the filename. Default: NULL}

\item{nrow}{number of rows for panels on the page. Default: 2}

\item{ncol}{number of columns for panels on the page. Default: 2}

\item{plot}{Show the plot? Default: F}

\item{labels}{Panel labels. Default: LETTERS}

\item{max.list.length}{Max number of panels (per page). Default: 16}

\item{extension}{file extension}

\item{scale}{Scaling factor of the canvas. Default: 1}

\item{h}{Height of the plot. Default: wA4 * scale}

\item{w}{Width of the plot. Default: hA4 * scale}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot up to 6 panels (3-by-1) on vertically standing A4 page.
}
\examples{
# qA4_grid_plot()
}
