% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qA4_grid_plot}
\alias{qA4_grid_plot}
\title{qA4_grid_plot}
\usage{
qA4_grid_plot(
  plot_list,
  suffix = NULL,
  plotname = sppp(substitute(plot_list), suffix),
  plot = F,
  nrow = 3,
  ncol = 2,
  extension = c("pdf", "png")[2],
  h = hA4 * scale,
  w = wA4 * scale,
  scale = 1,
  labels = LETTERS[1:length(plot_list)],
  ...
)
}
\arguments{
\item{plot_list}{A list of ggplot objects, each of which is one panel.}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{plotname}{Plot name, Default: F}

\item{nrow}{number of rows for panels on the page, Default: 2}

\item{ncol}{number of columns for panels on the page, Default: 2}

\item{extension}{file extension}

\item{h}{height of the plot, Default: wA4 * scale}

\item{w}{width of the plot, Default: hA4 * scale}

\item{scale}{Scaling factor of the canvas, Default: 1}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot up to 6 panels (3-by-1) on vertically standing A4 page.
}
\examples{
# qA4_grid_plot()
}
