% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot.df}
\alias{qbarplot.df}
\title{qbarplot.df - Barplot for tibbles or dataframes}
\usage{
qbarplot.df(
  df,
  x = colnames(df)[1],
  y = colnames(df)[2],
  plotname = FixPlotName(substitute(df)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  fill = colnames(df)[3],
  color = 1,
  palette_use = getOption("gg.palette_use", "jco"),
  label = NULL,
  plot = TRUE,
  save = TRUE,
  ext = "png",
  also.pdf = FALSE,
  save.obj = getOption("gg.save.obj", FALSE),
  scale = TRUE,
  position = if (scale) "fill" else "dodge",
  xlab.angle = 45,
  xlab = NULL,
  logY = FALSE,
  annotation_logticks_Y = logY,
  hline = FALSE,
  filtercol = 1,
  legend.position = "none",
  max.names = 50,
  limitsize = FALSE,
  max.categ = 10,
  grid = "y",
  mdlink = getOption("gg.mdlink", FALSE),
  w = qqqAxisLength(df),
  h = 5,
  ...
)
}
\arguments{
\item{df}{The variable to plot.}

\item{x}{Colname to split along X axis. Default: \code{colnames(df)[1]}.}

\item{y}{Colname to count along y axis. Default: \code{colnames(df)[2]}.
If omitted and that column is not numeric, a single numeric column in \code{df} is auto-detected.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{fill}{Color (split) by along Y. Default: \code{colnames(df)[3]}.}

\item{color}{Color (split) by along Y.}

\item{palette_use}{Color palette to use. Either a \code{ggpubr::get_palette} palette or a custom vector of colors. Default: 'jco'.}

\item{label}{Label text.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{scale}{Scale the Y axis to 100\%.}

\item{position}{Position adjustment of the bars. If \code{scale} is TRUE, position is set to "fill",
otherwise to "dodge". Can also be "stack".}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{xlab}{X-axis label. Default: NULL.}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{hline}{Draw a horizontal line on the plot.}

\item{filtercol}{Color bars below/above the threshold with red/green. Define the direction by -1 or 1. Takes effect if "*line" is defined.}

\item{legend.position}{Character indicating the position of the legend. Default is is 'none' to hide the legend.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{Limit size.}

\item{max.categ}{Maximum number of categories to show on the plot. Default is 10.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{mdlink}{Insert .pdf and .png image links to a markdown report file, setup by
\code{MarkdownReports::setup_MarkdownReports()}, located at \code{path_of_report}. Default: FALSE}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Draw and save a barplot for tibbles or dataframes
}
\examples{
my_tibble <- tibble::tibble(
  Column_1 = c("A", "A", "A", "B", "C", "C"),
  Column_2 = c("X", "Y", "Y", "Z", "X", "Z")
)
freq_table <- my_tibble |> dplyr::count(Column_1, Column_2)
qbarplot.df(freq_table, y = "n", fill = "Column_2")

}
