% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot.df}
\alias{qbarplot.df}
\title{qbarplot.df - Barplot for tibbles or dataframes}
\usage{
qbarplot.df(
  df,
  x = colnames(df)[1],
  y = colnames(df)[2],
  fill = colnames(df)[3],
  color = 1,
  label = NULL,
  also.pdf = FALSE,
  save.obj = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  plotname = FixPlotName(substitute(df)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  scale = TRUE,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  hline = FALSE,
  filtercol = 1,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  xlab.angle = 45,
  xlab = NULL,
  logY = FALSE,
  annotation_logticks_Y = logY,
  hide.legend = TRUE,
  max.names = 50,
  limitsize = FALSE,
  grid = "y",
  max.categ = 10,
  w = qqqAxisLength(df),
  h = 5,
  ...
)
}
\arguments{
\item{df}{The variable to plot.}

\item{x}{Colname to split along X axis. Default: \code{colnames(df)[1]}.}

\item{y}{Colname to count along y axis. Default: \code{colnames(df)[2]}.}

\item{fill}{Color (split) by along Y. Default: \code{colnames(df)[3]}.}

\item{color}{Color (split) by along Y.}

\item{label}{label}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{ext}{File extension (.pdf / .png).}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{scale}{Scale the Y axis to 100\%.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{hline}{Draw a horizontal line on the plot.}

\item{filtercol}{Color bars below/above the threshold with red/green. Define the direction by -1 or 1. Takes effect if "*line" is defined.}

\item{palette_use}{GGpubr color palette to use.}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{xlab}{X-axis label. Default: NULL.}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{hide.legend}{Hide legend.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{Limit size.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{max.categ}{Maximum number of categories to show on the plot. Default is 10.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Draw and save a barplot for tibbles or dataframes
}
\examples{
my_tibble <- tibble(
  Column_1 = c("A", "A", "A", "B", "C", "C"),
  Column_2 = c("X", "Y", "Y", "Z", "X", "Z")
)
freq_table <- my_tibble |> count(Column_1, Column_2)
qbarplot.df(freq_table)

}
