% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot}
\alias{qbarplot}
\title{Quickly draw and save a bar plot (png, pdf, ggobj.qs)}
\usage{
qbarplot(
  vec,
  also.pdf = FALSE,
  save = get0("b.save.plots.ggExpress", ifnotfound = TRUE),
  save.obj = get0("b.save.ggobj.ggExpress", ifnotfound = TRUE),
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  plot = TRUE,
  plotname = FixPlotName(substitute(vec)),
  subtitle = paste("Median:", iround(median(vec))),
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  mdlink = get0("b.mdlink", ifnotfound = FALSE),
  hline = FALSE,
  filtercol = 1,
  filtercol_default = TRUE,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  col = 1,
  xlab = "",
  xlab.angle = 45,
  logY = FALSE,
  ylim = c(0, iround(1.1 * as.numeric(max(vec, na.rm = TRUE)))),
  annotation_logticks_Y = logY,
  label = NULL,
  hide.legend = TRUE,
  legend.title = NULL,
  max.names = 100,
  limitsize = FALSE,
  grid = "y",
  ylab = NULL,
  w = qqqAxisLength(vec, factor = 0.25),
  h = 5,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save}{Save the plot into a file.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{hline}{Draw a horizontal line on the plot.}

\item{filtercol}{Color bars below/above the threshold with red/green. Define the direction by
-1 or 1. Takes effect if "*line" is defined.}

\item{filtercol_default}{Use default red/green colors for filtering. If FALSE, use the
first two colors of the palette. Default: TRUE, \code{c("#008B45FF", "#EE0000FF")}}

\item{palette_use}{Color palette to use. Either a \code{ggpubr::get_palette} palette or a custom vector of colors.}

\item{col}{The fill color of the bars. Default: 1st color of the palette.}

\item{xlab}{X-axis label. Default: "".}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{logY}{Make Y axis log10-scale.}

\item{ylim}{Y-axis limit values. Default: c(0, 1.1*max(vec))}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{label}{Label bars with their values, or by a custom string vector. Default: NULL (no labels).}

\item{hide.legend}{Hide legend. Default: TRUE.}

\item{legend.title}{Custom legend title. Provide a string.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{Limit size.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{ylab}{Y-axis label. Default: NULL.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\value{
It returns a ggplot object if \code{plot} is TRUE.
}
\description{
This all-in-one function draws, annotates, displays and saves a bar plot of a
distribution provided as a numeric vector. It is a wrapper around \code{ggpubr::ggbarplot()},
with the automation of many features. All \code{ggpubr} parameters can be accessed through
the \code{...} argument. It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
weight3 <- runif(12)
qbarplot(weight3, filtercol = -1, hline = .5)
qbarplot(weight3, filtercol = 1, hline = .5)
}
