% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot}
\alias{qbarplot}
\title{Barplot}
\usage{
qbarplot(
  vec,
  ext = "pdf",
  plot = TRUE,
  suffix = NULL,
  plotname = FixPlotName(kpp(substitute(vec), suffix)),
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = F),
  hline = F,
  filtercol = 1,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  col = as.character(1:3)[1],
  xlab.angle = 90,
  xlab = F,
  logY = FALSE,
  annotation_logticks_Y = logY,
  label = NULL,
  hide.legend = TRUE,
  max.names = 50,
  limitsize = FALSE,
  w = qqqAxisLength(vec),
  h = 5,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The name of the file and title of the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{hline}{Draw a horizontal line on the plot.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by -1 or 1. Takes effect if "*line" is defined.}

\item{palette_use}{GGpubr Color palette to use.}

\item{col}{Color of the plot.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{xlab}{X-axis label.}

\item{logY}{Make Y axis log10-scale.}

\item{label}{label}

\item{hide.legend}{hide legend}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{limitsize}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Barplot
}
\examples{
weight3 <- runif (12);
qbarplot(weight3, filtercol = -1, hline = .5);
qbarplot(weight3, filtercol = 1, hline = .5)
}
