% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot}
\alias{qbarplot}
\title{qbarplot}
\usage{
qbarplot(
  vec,
  also.pdf = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(def = "png"),
  plot = TRUE,
  suffix = NULL,
  plotname = FixPlotName(substitute(vec), suffix),
  filename = FALSE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  hline = FALSE,
  filtercol = 1,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  col = as.character(1:3)[1],
  xlab.angle = 45,
  xlab = "",
  logY = FALSE,
  ylim = c(0, 1.1 * as.numeric(max(vec, na.rm = TRUE))),
  annotation_logticks_Y = logY,
  label = NULL,
  hide.legend = TRUE,
  max.names = 50,
  limitsize = FALSE,
  grid = "y",
  w = qqqAxisLength(vec),
  h = 5,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{also.pdf}{Save plot in both png and pdf formats}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{hline}{Draw a horizontal line on the plot.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by -1 or 1. Takes effect if "*line" is defined.}

\item{palette_use}{GGpubr Color palette to use.}

\item{col}{Color of the plot.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{xlab}{X-axis label.}

\item{logY}{Make Y axis log10-scale.}

\item{ylim}{ylimit values}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{label}{label}

\item{hide.legend}{hide legend}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{limitsize}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Draw and save a barplot.
}
\examples{
weight3 <- runif (12);
qbarplot(weight3, filtercol = -1, hline = .5);
qbarplot(weight3, filtercol = 1, hline = .5)
}
