% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qbarplot.stacked.from.wide.df}
\alias{qbarplot.stacked.from.wide.df}
\title{qbarplot.stacked.from.wide.df - Barplot for tibbles or dataframes}
\usage{
qbarplot.stacked.from.wide.df(
  df,
  x = "Samples",
  y = "Fraction",
  z = "Category",
  color = 1,
  label = NULL,
  also.pdf = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  plotname = FixPlotName(substitute(df)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  scale = TRUE,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  hline = FALSE,
  filtercol = 1,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  xlab.angle = 45,
  xlab = x,
  logY = FALSE,
  annotation_logticks_Y = logY,
  hide.legend = FALSE,
  max.names = 50,
  limitsize = FALSE,
  grid = "y",
  max.categ = 10,
  top = NULL,
  w = qqqAxisLength(df, factor = 0.7),
  h = 5,
  ...
)
}
\arguments{
\item{df}{The variable to plot.}

\item{x}{Colname to split along X axis. Def colnames(df)\link{1}}

\item{y}{Colname to count along y axis. Def colnames(df)\link{3}}

\item{color}{Color (split) by along Y.}

\item{label}{label}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{ext}{File extension (.pdf / .png).}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{hline}{Draw a horizontal line on the plot.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by -1 or 1. Takes effect if "*line" is defined.}

\item{palette_use}{GGpubr Color palette to use.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{xlab}{X-axis label.}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{hide.legend}{hide legend}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{limitsize}{limitsize}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}

\item{fill}{Color (split) by along Y.}
}
\description{
Draw and save a stacked barplot for each row of a dataframe.
}
\examples{
# Example of wide-format data for stacked bar plot
df.SingletSplit <- tibble::tibble(
  doublet = c(0.027886224, 0.007699141, 0.003704390, 0.003205128),
  singlet = c(0.9686280, 0.9872668, 0.9925912, 0.9119822),
  unassigned = c(0.0034857780, 0.0050340539, 0.0037043897, 0.0848126233)
)
rownames(df.SingletSplit) <- c("sc06.692", "sc06.693", "sc08.325", "sc08.327")

# Create the stacked bar plot using the new qbarplot.df2 function
qbarplot.stacked.from.wide.df(df.SingletSplit)

}
