% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qboxplot}
\alias{qboxplot}
\title{qboxplot}
\usage{
qboxplot(
  df_XYcol_or_list,
  x = 1,
  y = 2,
  col = NULL,
  fill = "gold",
  plotname = FixPlotName(substitute(df_XYcol_or_list)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  outlier.shape = NULL,
  stat.test = TRUE,
  stat.method = NULL,
  stat.label.y.npc = "top",
  stat.label.x = NULL,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  hide.legend = FALSE,
  also.pdf = TRUE,
  save.obj = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  ylab = NULL,
  logY = FALSE,
  annotation_logticks_Y = logY,
  xlab.angle = 90,
  hline = FALSE,
  vline = FALSE,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  grid = "y",
  max.categ = 100,
  add = NULL,
  w = qqqAxisLength(df_XYcol_or_list),
  h = 6,
  ...
)
}
\arguments{
\item{df_XYcol_or_list}{Data, as 2 column data frame, where col.1 is X axis, alternatively a uniquely named list ov values.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{outlier.shape}{outlier shape. NA to hide.}

\item{stat.test}{Do a statistical test?}

\item{stat.method}{stat method. NULL for default}

\item{stat.label.y.npc}{stat label y position}

\item{stat.label.x}{stat label x position}

\item{palette_use}{GGpubr Color palette to use.}

\item{hide.legend}{hide legend}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{ext}{File extension (.pdf / .png).}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{hline}{Draw a horizontal line on the plot.}

\item{vline}{Draw a vertical line on the plot.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{max.categ}{The maximum allowed number of unique categories.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Draw and save a boxplot
}
\examples{
data("ToothGrowth")
ToothLen.by.Dose <- ToothGrowth[, c("dose", "len")]
qboxplot(ToothLen.by.Dose)
}
