% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qdensity}
\alias{qdensity}
\title{Quickly draw and save a density plot (png, pdf, ggobj.qs)}
\usage{
qdensity(
  vec,
  plotname = FixPlotName(substitute(vec)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  plot = TRUE,
  save = TRUE,
  ext = "png",
  also.pdf = FALSE,
  save.obj = getOption("gg.save.obj", FALSE),
  xlab = FALSE,
  xlab.angle = 90,
  logX = FALSE,
  logY = FALSE,
  palette_use = getOption("gg.palette_use", "jco"),
  legend.position = "none",
  max.names = 50,
  grid = FALSE,
  mdlink = getOption("gg.mdlink", FALSE),
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{xlab}{X-axis label. Default: FALSE.}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{palette_use}{Color palette to use. Either a \code{ggpubr::get_palette} palette or a custom vector of colors. Default: 'jco'.}

\item{legend.position}{Character indicating the position of the legend. Default is is 'none' to hide the legend.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{mdlink}{Insert .pdf and .png image links to a markdown report file, setup by
\code{MarkdownReports::setup_MarkdownReports()}, located at \code{path_of_report}. Default: FALSE}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
This all-in-one function draws, annotates, displays and saves a density plot of a
distribution provided as a numeric vector. It is a wrapper around \code{ggpubr::ggdensity()},
with the automation of many features. All \code{ggpubr} parameters can be accessed through
the \code{...} argument.  It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
weight <- rnorm(1000)
qdensity(weight)
}
