% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qdensity}
\alias{qdensity}
\title{Quickly draw and save a density plot (png, pdf, ggobj.qs)}
\usage{
qdensity(
  vec,
  also.pdf = FALSE,
  save.obj = FALSE,
  ext = "png",
  xlab = FALSE,
  plot = TRUE,
  xlab.angle = 90,
  plotname = FixPlotName(substitute(vec)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  save = TRUE,
  mdlink = get0("b.mdlink", ifnotfound = FALSE),
  logX = FALSE,
  logY = FALSE,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  hide.legend = TRUE,
  max.names = 50,
  grid = FALSE,
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{ext}{File extension (.pdf / .png).}

\item{xlab}{X-axis label. Default: FALSE.}

\item{plot}{Display the plot.}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{palette_use}{GGpubr color palette to use.}

\item{hide.legend}{Hide legend.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
This all-in-one function draws, annotates, displays and saves a density plot of a
distribution provided as a numeric vector. It is a wrapper around \code{ggpubr::ggdensity()},
with the automation of many features. All \code{ggpubr} parameters can be accessed through
the \code{...} argument.  It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
weight <- rnorm(1000)
qdensity(weight)
}
