% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qdensity}
\alias{qdensity}
\title{qdensity}
\usage{
qdensity(
  vec,
  ext = MarkdownHelpers::unless.specified("b.def.ext", def = "pdf"),
  xlab = FALSE,
  plot = TRUE,
  suffix = NULL,
  plotname = FixPlotName(substitute(vec), suffix),
  filename = FALSE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  logX = FALSE,
  logY = FALSE,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  xlab.angle = 90,
  hide.legend = TRUE,
  max.names = 50,
  grid = F,
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{xlab}{X-axis label.}

\item{plot}{Display the plot.}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{palette_use}{GGpubr Color palette to use.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{hide.legend}{hide legend}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Draw and save a density plot.
}
\examples{
weight <- rnorm(1000); qdensity(weight)
}
