% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qdensity}
\alias{qdensity}
\title{Density plot}
\usage{
qdensity(
  vec,
  ext = "pdf",
  xlab = F,
  plot = TRUE,
  suffix = NULL,
  plotname = sppp(substitute(vec), suffix),
  save = TRUE,
  mdlink = FALSE,
  logX = F,
  logY = F,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  max.names = 50,
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{xlab}{X-axis label.}

\item{plot}{Display the plot.}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The name of the file and title of the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{palette_use}{GGpubr Color palette to use.}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Density plot
}
\examples{
weight <- rnorm(1000); qdensity(weight)
}
