% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qheatmap}
\alias{qheatmap}
\title{Quick Heatmap Plot}
\usage{
qheatmap(
  data_matrix,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  plotname = paste("Heatmap of", substitute(data_matrix)),
  subtitle = NULL,
  caption = NULL,
  suffix = NULL,
  filename = NULL,
  also.pdf = FALSE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  color = colorRampPalette(c("#0073c2", "white", "#efc000"))(100),
  legendName = "Intensity",
  scale = "row",
  cluster_rows = F,
  cluster_cols = F,
  annotation_rows = NULL,
  annotation_cols = NULL,
  annotation_color = NULL,
  w = 10,
  h = 8,
  xlab = "Global X Label",
  ylab = "Global Y Label",
  ...
)
}
\arguments{
\item{data_matrix}{The data matrix to be plotted; must be either a matrix or a data frame.
Default: None (must be provided).}

\item{ext}{The file extension for the saved plot image, adjusted based on \code{also.pdf}.
Default: "png".}

\item{plotname}{The base name for the plot file. Default: "heatmap".}

\item{subtitle}{The subtitle for the heatmap. Default: "NULL" (indicating no subtitle).}

\item{caption}{The caption for the heatmap. Default: "mousse".}

\item{suffix}{An optional suffix to append to the plot filename. Default: NULL.}

\item{filename}{The specific filename to save the plot as; if NULL, a name is generated based on
\code{plotname} and other parameters. Default: NULL.}

\item{also.pdf}{Logical; indicates whether to also save the plot as a PDF. Default: FALSE.}

\item{save}{Logical; indicates whether to save the plot to file. Default: TRUE.}

\item{mdlink}{Logical; if TRUE and \code{save} is also TRUE, generates a markdown link for the saved
plot image. Default: FALSE, using \code{MarkdownHelpers::unless.specified}.}

\item{color}{The color palette for the heatmap values, using a blue-white-yellow gradient.
Default: colorRampPalette(c( "#0073c2","white","#efc000"))(100).}

\item{legendName}{The title of the heatmap legend. Default: "Intensity".}

\item{scale}{Character; indicates how to scale the data ("none", "row", or "column").
Default: "row".}

\item{cluster_rows}{Logical; indicates whether to cluster rows. Default: FALSE.}

\item{cluster_cols}{Logical; indicates whether to cluster columns. Default: FALSE.}

\item{annotation_rows}{Data frame with row annotations. Default: NULL.}

\item{annotation_cols}{Data frame with column annotations. Default: NULL.}

\item{annotation_color}{List of colors for the annotations. Default: NULL.}

\item{w}{Width of the saved plot image, in inches. Default: 10.}

\item{h}{Height of the saved plot image, in inches. Default: 8.}

\item{xlab}{The global label for the x-axis; this might not display due to limitations.
Default: "Global X Label".}

\item{ylab}{The global label for the y-axis; this might not display due to limitations.
Default: "Global Y Label".}

\item{...}{Additional arguments passed to \code{ggheatmap}.}

\item{title}{The main title for the heatmap. Default: "Heatmap".}
}
\value{
A \code{ggplot} object representing the heatmap, conditionally modified based on the provided
parameters.
}
\description{
Generates a heatmap plot using the \code{ggheatmap} package with features such as automatic
file saving and markdown link generation. This function simplifies the process of creating and
customizing heatmaps, with support for clustering, annotations, and scaling.
}
\examples{
\dontrun{
  set.seed(123)
  data_matrix <- matrix(rnorm(100), ncol = 10)
  qheatmap(data_matrix, title = "My Heatmap", save = FALSE)
}

}
