% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qhistogram}
\alias{qhistogram}
\title{Quickly draw and save a histogram (png, pdf, ggobj.qs)}
\usage{
qhistogram(
  vec,
  plotname = FixPlotName(substitute(vec)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  plot = TRUE,
  save = TRUE,
  also.pdf = FALSE,
  ext = "png",
  save.obj = getOption("gg.save.obj", FALSE),
  xlab = FALSE,
  logX = FALSE,
  logY = FALSE,
  annotation_logticks_X = logX,
  annotation_logticks_Y = logY,
  vline = FALSE,
  filtercol = 0,
  add = "median",
  palette_use = getOption("gg.palette_use", "jco"),
  col = as.character(1:3)[1],
  xlab.angle = 90,
  legend.position = "none",
  max.names = 50,
  grid = "y",
  mdlink = getOption("gg.mdlink", FALSE),
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{A numeric vector for which the histogram is to be plotted.}

\item{plotname}{Title of the plot and the name of the file (unless specified in \code{filename}). Default is parsed from \code{vec}.}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Optional filename for the saved plot. Default is parsed from \code{plotname}.}

\item{plot}{Logical indicating whether to display the plot. Default is TRUE.}

\item{save}{Logical indicating whether to save the plot into a file. Default is TRUE.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{ext}{File extension for the saved plot. Either '.pdf' or '.png'. Default is '.png'.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{xlab}{Label for the X-axis. By default, it uses the plot name.}

\item{logX}{Logical indicating whether to make X axis on log10 scale. Default is FALSE.}

\item{logY}{Logical indicating whether to make Y axis on log10 scale. Default is FALSE.}

\item{annotation_logticks_X}{Logical indicating whether to add annotation logticks on X-axis. Default follows the value of \code{logX}.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{vline}{Numeric value at which to draw a vertical line on the plot. Default is FALSE (no line).}

\item{filtercol}{Numeric value indicating the direction to color bars above/below the threshold. Default is 0 (no color change).}

\item{add}{Character defining the type of plot annotations to add. Default is 'median'.}

\item{palette_use}{Color palette to use. Either a \code{ggpubr::get_palette} palette or a custom vector of colors. Default: 'jco'.}

\item{col}{Color of the plot. Default is '1'.}

\item{xlab.angle}{Angle to rotate X-axis labels. Default is 90 degrees.}

\item{legend.position}{Character indicating the position of the legend. Default is is 'none' to hide the legend.}

\item{max.names}{Maximum number of names to show on the axis. Default is 50.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{mdlink}{Insert .pdf and .png image links to a markdown report file, setup by
\code{MarkdownReports::setup_MarkdownReports()}, located at \code{path_of_report}. Default: FALSE}

\item{w}{Width of the plot. Default is 5.}

\item{h}{Height of the plot. Default is the same as width.}

\item{...}{Additional parameters for the corresponding plotting function.}
}
\value{
It returns a ggplot object if \code{plot} is TRUE.
}
\description{
This all-in-one function draws, annotates, displays and saves a histogram of a
distribution provided as a numeric vector. It is a wrapper around \code{ggpubr::gghistogram()},
with the automation of many features. All \code{ggpubr} parameters can be accessed through
the \code{...} argument. It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
\dontrun{
weight <- rnorm(1000)
qhistogram(vec = weight)
qhistogram(vec = weight, vline = 2, filtercol = -1)
}

}
