% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qhistogram}
\alias{qhistogram}
\title{Quick Histogram Plotting}
\usage{
qhistogram(
  vec,
  also.pdf = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  xlab = FALSE,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  plotname = FixPlotName(substitute(vec)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  logX = FALSE,
  logY = FALSE,
  annotation_logticks_X = logX,
  annotation_logticks_Y = logY,
  vline = FALSE,
  filtercol = 0,
  add = "median",
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  col = as.character(1:3)[1],
  xlab.angle = 90,
  hide.legend = TRUE,
  max.names = 50,
  grid = "y",
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{A numeric vector for which the histogram is to be plotted.}

\item{also.pdf}{Save plot in both png and pdf formats}

\item{ext}{File extension for the saved plot. Either '.pdf' or '.png'. Default is '.pdf'.}

\item{xlab}{Label for the X-axis. By default, it uses the plot name.}

\item{plot}{Logical indicating whether to display the plot. Default is TRUE.}

\item{save}{Logical indicating whether to save the plot into a file. Default is TRUE.}

\item{mdlink}{Logical indicating whether to insert .pdf and .png image links in the markdown report, set by "path_of_report". Default is FALSE.}

\item{plotname}{Title of the plot and the name of the file (unless specified in \code{filename}). Default is parsed from \code{vec}.}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix}

\item{filename}{Optional filename for the saved plot. Default is parsed from \code{plotname}.}

\item{logX}{Logical indicating whether to make X axis on log10 scale. Default is FALSE.}

\item{logY}{Logical indicating whether to make Y axis on log10 scale. Default is FALSE.}

\item{annotation_logticks_X}{Logical indicating whether to add annotation logticks on X-axis. Default follows the value of \code{logX}.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{vline}{Numeric value at which to draw a vertical line on the plot. Default is FALSE (no line).}

\item{filtercol}{Numeric value indicating the direction to color bars above/below the threshold. Default is 0 (no color change).}

\item{add}{Character defining the type of plot annotations to add. Default is 'median'.}

\item{palette_use}{Color palette to use from GGpubr. Default is 'jco'.}

\item{col}{Color of the plot. Default is '1'.}

\item{xlab.angle}{Angle to rotate X-axis labels. Default is 90 degrees.}

\item{hide.legend}{Logical indicating whether to hide the legend. Default is TRUE.}

\item{max.names}{Maximum number of names to show on the axis. Default is 50.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{w}{Width of the plot. Default is 5.}

\item{h}{Height of the plot. Default is the same as width.}

\item{...}{Additional parameters for the corresponding plotting function.}
}
\value{
It returns a ggplot object if \code{plot} is TRUE.
}
\description{
This function generates a histogram and saves the plot for a given vector and offers several customizations.
}
\examples{
\dontrun{
weight <- rnorm(1000)
qhistogram(vec = weight)
qhistogram(vec = weight, vline = 2, filtercol = -1)
}

}
