% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qhistogram}
\alias{qhistogram}
\title{Histogram}
\usage{
qhistogram(
  vec,
  ext = "pdf",
  xlab = F,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = F),
  suffix = NULL,
  plotname = sppp(substitute(vec), suffix),
  logX = F,
  logY = F,
  vline = F,
  filtercol = 0,
  add = "median",
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  col = as.character(1:3)[1],
  xlab.angle = 90,
  hide.legend = TRUE,
  max.names = 50,
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{xlab}{X-axis label.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The name of the file and title of the plot.}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{vline}{Draw a vertical line on the plot.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by -1 or 1. Takes effect if "*line" is defined.}

\item{palette_use}{GGpubr Color palette to use.}

\item{col}{Color of the plot.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{hide.legend}{hide legend}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Histogram
}
\examples{
weight <- rnorm(1000); qhistogram(vec = weight); qhistogram(vec = weight, vline = 2, filtercol = -1)
}
