% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qMarkdownImageLink}
\alias{qMarkdownImageLink}
\title{qMarkdownImageLink}
\usage{
qMarkdownImageLink(file_name = "myplot.pdf")
}
\arguments{
\item{file_name}{File name (basename) of the image file to convert into a markdown image link.}
}
\description{
Insert Markdown image link to .md report
}
\examples{
qMarkdownImageLink(file_name = "myplot.pdf")
}
