% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qpie}
\alias{qpie}
\title{Quickly draw and save a pie chart (png, pdf, ggobj.qs)}
\usage{
qpie(
  vec,
  also.pdf = FALSE,
  save.obj = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  plotname = FixPlotName(substitute(vec)),
  filename = NULL,
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  caption.ext = TRUE,
  NoLegend = FALSE,
  LegendSide = TRUE,
  LegendTitle = "",
  pcdigits = 2,
  NamedSlices = FALSE,
  custom.order = FALSE,
  extended.canvas = TRUE,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  custom.margin = TRUE,
  max.categories = 100,
  max.names = 10,
  decr.order = TRUE,
  both_pc_and_value = FALSE,
  labels = "names",
  w = 7,
  h = 5,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{NoLegend}{No legend.}

\item{LegendSide}{Legend side.}

\item{LegendTitle}{Legend title.}

\item{pcdigits}{Number of digits for percentages.}

\item{NamedSlices}{Use named slices.}

\item{custom.order}{Custom order.}

\item{extended.canvas}{Make an extended canvas. Default: TRUE.}

\item{palette_use}{GGpubr color palette to use.}

\item{custom.margin}{Custom plot margin. Default: TRUE.}

\item{max.categories}{Maximum number of categories to be shown as a separate slice}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{decr.order}{Slices in the order of df. By default, they are ordered alphabetically in the plot.}

\item{both_pc_and_value}{Report both percentage and number.}

\item{labels}{Slice labels. Set to NULL to remove slice names.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
This all-in-one function draws, annotates, displays and saves a pie chart of a
distribution provided as a numeric table. It is a wrapper around \code{ggpubr::ggpie()},
with the automation of many features. All \code{ggpubr} parameters can be accessed through
the \code{...} argument. It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
xvec <- c("A" = 12, "B" = 29)
qpie(vec = xvec)
}
