% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qpie}
\alias{qpie}
\title{Pie chart}
\usage{
qpie(
  vec = Network.Size,
  ext = "pdf",
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = F),
  suffix = NULL,
  plotname = sppp(substitute(vec), suffix),
  LegendSide = TRUE,
  LegendTitle = as.character(substitute(vec)),
  NoLegend = F,
  pcdigits = 2,
  NamedSlices = FALSE,
  custom.order = FALSE,
  extended.canvas = TRUE,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  custom.margin = T,
  max.categories = 100,
  max.names = 10,
  decr.order = TRUE,
  both_pc_and_value = TRUE,
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The name of the file and title of the plot.}

\item{LegendSide}{LegendSide}

\item{LegendTitle}{LegendTitle}

\item{NoLegend}{NoLegend}

\item{pcdigits}{pcdigits}

\item{NamedSlices}{NamedSlices}

\item{custom.order}{custom.order}

\item{extended.canvas}{Make an extended canvas, default: T}

\item{palette_use}{GGpubr Color palette to use.}

\item{custom.margin}{custom plot margin, default: T}

\item{max.categories}{Maximum number of categories to be shown as a seprate slice}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{decr.order}{Slices in the order of df. By default would ordered alphabetically in the plot.}

\item{both_pc_and_value}{Report both percentage AND number.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Pie chart
}
\examples{
xvec <- c("A"=12, "B"=29); qpie(vec = xvec)
}
