% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qpie}
\alias{qpie}
\title{Pie chart}
\usage{
qpie(
  vec,
  ext = "pdf",
  plot = TRUE,
  save = TRUE,
  mdlink = FALSE,
  suffix = NULL,
  plotname = qqqParsePlotname(vec, suffix),
  LegendSide = T,
  LegendTitle = as.character(substitute(vec)),
  NoLegend = F,
  pcdigits = 2,
  NamedSlices = F,
  custom.order = F,
  color.palette = "jco",
  max.names = 50,
  w = 5,
  h = w,
  ...
)
}
\arguments{
\item{vec}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The name of the file and title of the plot.}

\item{LegendSide}{LegendSide}

\item{LegendTitle}{LegendTitle}

\item{NoLegend}{NoLegend}

\item{pcdigits}{pcdigits}

\item{NamedSlices}{NamedSlices}

\item{custom.order}{custom.order}

\item{color.palette}{color.palette}

\item{max.names}{The maximum number of names still to be shown on the axis.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Pie chart
}
\examples{
xvec <- c("A"=12, "B"=29); qpie(vec = xvec)
}
