% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqAxisLength}
\alias{qqqAxisLength}
\title{qqqAxisLength}
\usage{
qqqAxisLength(vec = 1:20, minLength = 6, factor = 0.4)
}
\arguments{
\item{vec}{The variable to plot.}

\item{minLength}{minLength}

\item{factor}{Length adjustment factor.}
}
\description{
Define Axis Length
}
\examples{
qqqAxisLength()
}
