% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqAxisLength}
\alias{qqqAxisLength}
\title{Define Axis Length}
\usage{
qqqAxisLength(vec = 1:20, minLength = 6)
}
\arguments{
\item{vec}{The variable to plot.}

\item{minLength}{minLength}
}
\description{
Define Axis Length
}
\examples{
qqqAxisLength()
}
