% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqCovert.tbl2vec}
\alias{qqqCovert.tbl2vec}
\title{qqqCovert.tbl2vec}
\usage{
qqqCovert.tbl2vec(tibble.input, name.column = 1, value.column = 2)
}
\arguments{
\item{tibble.input}{tibble.input}

\item{name.column}{name.column}

\item{value.column}{value.column}
}
\description{
Covert a table to a named vector.
}
\examples{
a=1:5; x= tibble::tibble(a, a * 2); qqqCovert.tbl2vec(x)
}
