% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqNamed.Vec.2.Tbl}
\alias{qqqNamed.Vec.2.Tbl}
\title{qqqNamed.Vec.2.Tbl}
\usage{
qqqNamed.Vec.2.Tbl(
  namedVec = 1:14,
  verbose = FALSE,
  strip.too.many.names = TRUE,
  thr = 50
)
}
\arguments{
\item{namedVec}{namedVec}

\item{verbose}{verbose}

\item{strip.too.many.names}{strip.too.many.names}

\item{thr}{thr}
}
\description{
Covert a named vector to a table.
}
\examples{
qqqNamed.Vec.2.Tbl(namedVec = c("A" = 2, "B" = 29))
}
