% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqqParsePlotname}
\alias{qqqParsePlotname}
\title{qqqParsePlotname}
\usage{
qqqParsePlotname(string = "sadsad", suffix_tag = NULL)
}
\arguments{
\item{string}{string}

\item{suffix_tag}{suffix_tag}
}
\description{
Parse Plotname from variable name.
}
\examples{
qqqParsePlotname()
}
