% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqSave}
\alias{qqSave}
\title{qqSave}
\usage{
qqSave(
  ggobj,
  title = FALSE,
  fname = FALSE,
  suffix = NULL,
  ext = "png",
  subdir = FALSE,
  png.subdir = FALSE,
  png.dir.name = "pdf",
  also.pdf = getOption("gg.save.pdf", T),
  pdf.subdir = FALSE,
  pdf.dir.name = "pdf",
  save.obj = getOption("gg.save.obj", F),
  obj.subdir = FALSE,
  obj.dir.name = "ggobj",
  save.meta.info = getOption("gg.save.meta", F),
  meta.info.custom = "",
  bgcol = "white",
  max.obj.size = 5e+06,
  w = 6,
  h = w,
  plot_on_page = list(FALSE, "A4p", "A4l", "A5p", "A5l")[[1]],
  ...
)
}
\arguments{
\item{ggobj}{Plot as ggplot object.}

\item{title}{title field for pdf file (saved into file metadata)}

\item{fname}{Manual filename}

\item{suffix}{A suffix added to the filename. Default: NULL.}

\item{ext}{File extension. Default: "png". Ignored atm, but needed for ggX... functions to not fail.}

\item{subdir}{Save all files within a subdirectory. The subdirectory is named after the string
provided here, and will be created using MarkdownReports. Default: FALSE.}

\item{png.subdir}{Save png files into a subdirectory. Default: FALSE.}

\item{png.dir.name}{Name of the png subdirectory. Default: "png".}

\item{also.pdf}{Save plot in both png and pdf formats. Default: FALSE.}

\item{pdf.subdir}{Save pdf files into a subdirectory. Default: FALSE.}

\item{pdf.dir.name}{Name of the pdf subdirectory. Default: "pdf".}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{obj.subdir}{Save ggplot objects into a subdirectory. Default: FALSE.}

\item{obj.dir.name}{Name of the ggplot object subdirectory. Default: "ggobj".#'}

\item{save.meta.info}{Save metadata (script of origin, original location) as a sidecar
metadata file "fname.meta.txt" ? Default: TRUE.}

\item{meta.info.custom}{Custom text to add to the metadata file. Default: "".}

\item{bgcol}{Plot background color. Default: "white".}

\item{max.obj.size}{Maximum allowed size of the ggplot object to be saved (in bytes). Default: 5e+06 (5 MB).}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{plot_on_page}{Use standard page sizes for plots? Options are:
FALSE, "A4p", "A4l", "A5p", "A5l". Overwrites w and h.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Quick-Save ggplot objects to file with automatic file naming.
Default format: png, optionally also saves a .pdf and/or .qs ggplot object.
}
\examples{
xplot <- ggplot2::qplot(12)
qqSave(ggobj = xplot)
qqSave(ggobj = xplot, ext = "pdf")

}
