% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqSave}
\alias{qqSave}
\title{qqSave}
\usage{
qqSave(
  ggobj,
  ext = MarkdownHelpers::unless.specified("b.def.ext", def = "png"),
  also.pdf = FALSE,
  save.obj = FALSE,
  max.obj.size = 5e+06,
  bgcol = "white",
  page = c(F, "A4p", "A4l", "A5p", "A5l")[1],
  title = FALSE,
  fname = FALSE,
  suffix = NULL,
  w = 4,
  h = w,
  ...
)
}
\arguments{
\item{ggobj}{Plot as ggplot object.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{Save plot in both png and pdf formats. Default: FALSE.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{max.obj.size}{Maximum allowed size of the ggplot object to be saved (in bytes). Default: 5e+06 (5 MB).}

\item{bgcol}{Plot background color. Default: "white".}

\item{page}{Set page size to a predefined value, eg. "A4".
Default: \code{c(F, "A4p", "A4l", "A5p", "A5l")[1]}}

\item{title}{title field for pdf file (saved into file metadata)}

\item{fname}{Manual filename}

\item{suffix}{A suffix added to the filename. Default: NULL.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Quick-Save ggplot objects to file (png / pdf) with automatic naming and as .qs object.
}
\examples{
xplot <- ggplot2::qplot(12)
qqSave(ggobj = xplot)
qqSave(ggobj = xplot, ext = "pdf")

}
