% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.auxiliary.functions.R
\name{qqSave}
\alias{qqSave}
\title{Quick-Save ggplot objects}
\usage{
qqSave(
  ggobj,
  w = 4,
  h = w,
  ext = c("png", "pdf")[1],
  also.pdf = FALSE,
  page = c(F, "A4p", "A4l", "A5p", "A5l")[1],
  title = F,
  fname = F,
  suffix = NULL,
  ...
)
}
\arguments{
\item{ggobj}{ggobj}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{also.pdf}

\item{page}{page}

\item{title}{title}

\item{fname}{fname}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Quick-Save ggplot objects
}
\examples{
xplot <- ggplot2::qplot(12); qqSave(ggobj = xplot); qqSave(ggobj = xplot, ext = "pdf")
}
