% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqSave}
\alias{qqSave}
\title{qqSave}
\usage{
qqSave(
  ggobj,
  title = FALSE,
  fname = FALSE,
  suffix = NULL,
  ext = "png",
  also.pdf = FALSE,
  pdf.subdir = FALSE,
  pdf.dir.name = "pdf",
  save.obj = getOption("gg.save.obj", F),
  obj.subdir = FALSE,
  obj.dir.name = "ggobj",
  bgcol = "white",
  max.obj.size = 5e+06,
  w = 6,
  h = w,
  plot_on_page = list(FALSE, "A4p", "A4l", "A5p", "A5l")[[1]],
  ...
)
}
\arguments{
\item{ggobj}{Plot as ggplot object.}

\item{title}{title field for pdf file (saved into file metadata)}

\item{fname}{Manual filename}

\item{suffix}{A suffix added to the filename. Default: NULL.}

\item{also.pdf}{Save plot in both png and pdf formats. Default: FALSE.}

\item{pdf.subdir}{Save pdf files into a subdirectory. Default: FALSE.}

\item{pdf.dir.name}{Name of the pdf subdirectory. Default: "pdf".}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{obj.subdir}{Save ggplot objects into a subdirectory. Default: FALSE.}

\item{obj.dir.name}{Name of the ggplot object subdirectory. Default: "ggobj".#'}

\item{bgcol}{Plot background color. Default: "white".}

\item{max.obj.size}{Maximum allowed size of the ggplot object to be saved (in bytes). Default: 5e+06 (5 MB).}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{plot_on_page}{Use standard page sizes for plots? Options are:
FALSE, "A4p", "A4l", "A5p", "A5l". Overwrites w and h.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Quick-Save ggplot objects to file with automatic file naming.
Default format: png, optionally also: .pdf and .qs ggplot object.
}
\examples{
xplot <- ggplot2::qplot(12)
qqSave(ggobj = xplot)
qqSave(ggobj = xplot, ext = "pdf")

}
