% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qqSave}
\alias{qqSave}
\title{qqSave}
\usage{
qqSave(
  ggobj,
  w = 4,
  h = w,
  ext = MarkdownHelpers::unless.specified("b.def.ext", def = "png"),
  also.pdf = FALSE,
  bgcol = "white",
  page = c(F, "A4p", "A4l", "A5p", "A5l")[1],
  title = FALSE,
  fname = FALSE,
  suffix = NULL,
  ...
)
}
\arguments{
\item{ggobj}{Plot as ggplot object.}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{Save plot in both png and pdf formats}

\item{bgcol}{Plot background color}

\item{page}{page}

\item{title}{title field for pdf file (saved into file metadata)}

\item{fname}{fname}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Quick-Save ggplot objects
}
\examples{
xplot <- ggplot2::qplot(12); qqSave(ggobj = xplot); qqSave(ggobj = xplot, ext = "pdf")
}
