% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qscatter}
\alias{qscatter}
\title{Quickly draw and save a scatter plot (png, pdf, ggobj.qs)}
\usage{
qscatter(
  df_XYcol,
  x = 1,
  y = 2,
  plotname = FixPlotName(substitute(df_XYcol)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  plot = TRUE,
  save = TRUE,
  save.obj = FALSE,
  also.pdf = TRUE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  label = NULL,
  repel = TRUE,
  hide.legend = FALSE,
  col = c(NULL, 3)[1],
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  xlab = NULL,
  ylab = NULL,
  xlab.angle = 90,
  logX = FALSE,
  logY = FALSE,
  annotation_logticks_Y = logY,
  annotation_logticks_X = logX,
  xlim = NULL,
  ylim = NULL,
  isometric = FALSE,
  hline = FALSE,
  vline = FALSE,
  abline = FALSE,
  line.col = "darkgrey",
  line.width = 0.5,
  line.type = "dashed",
  add_contour_plot = FALSE,
  correlation_r2 = FALSE,
  w = 7,
  h = w,
  grid = "xy",
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  ...
)
}
\arguments{
\item{df_XYcol}{Data, as 2 column data frame, where col.1 is X axis.}

\item{x}{The index or name of the column to be plotted on the X axis. Default: \code{1}.}

\item{y}{The index or name of the column to be plotted on the Y axis. Default: \code{2}.}

\item{plotname}{The name of the file and title of the plot.}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{ext}{File extension (.pdf / .png).}

\item{label}{Point labels. Default: NULL.}

\item{repel}{Repel labels from each other. Default: TRUE.}

\item{hide.legend}{Hide legend.}

\item{col}{Color of the plot.}

\item{palette_use}{GGpubr color palette to use.}

\item{xlab}{X-axis label. Default: NULL.}

\item{ylab}{Y-axis label. Default: NULL.}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{annotation_logticks_X}{Logical indicating whether to add annotation logticks on X-axis. Default follows the value of \code{logX}.}

\item{xlim}{X-axis limit values. Default: NULL.}

\item{ylim}{Y-axis limit values. Default: NULL.}

\item{isometric}{Make both axes isometric (same scale).}

\item{hline}{Draw a horizontal line on the plot, yintercept or FALSE}

\item{vline}{Draw a vertical line on the plot, xintercept or FALSE.}

\item{abline}{Draw a sloped line on the plot. Set to FALSE, or \verb{intercept = abline[1], slope = abline[2]}.}

\item{line.col}{Color of the lines (vline, hline, abline). Default: "darkgrey".}

\item{line.width}{Width of the line (vline, hline, abline). Default: 0.5.}

\item{line.type}{Type of the line (vline, hline, abline). Default: "dashed".}

\item{add_contour_plot}{Add 2D contour plot. See: http://www.sthda.com/english/articles/32-r-graphics-essentials/131-plot-two-continuous-variables-scatter-graph-and-alternatives/#continuous-bivariate-distribution}

\item{correlation_r2}{Add a correlation value to the plot}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
This all-in-one function draws, annotates, displays and saves a scatter plot of
two variables provided as a 2-column data frame or matrix. It is a wrapper around
\code{ggpubr::ggscatter()}, with the automation of many features. All \code{ggpubr} parameters can be
accessed through the \code{...} argument. It can automatically save the plot as png (default) and/or
pdf files, and the ggplot object as a .qs file.
}
\examples{
dfx <- as.data.frame(cbind("AA" = rnorm(500), "BB" = rnorm(500)))
qscatter(dfx, suffix = "2D.gaussian")

}
