% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qscatter}
\alias{qscatter}
\title{Scatter plot}
\usage{
qscatter(
  df_XYcol,
  suffix = NULL,
  plotname = sppp(substitute(df_XYcol), suffix),
  col = c(NULL, 3)[1],
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  hide.legend = FALSE,
  ext = "png",
  also.pdf = T,
  logX = F,
  logY = F,
  xlab.angle = 90,
  hline = F,
  vline = F,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = F),
  w = 7,
  h = w,
  ...
)
}
\arguments{
\item{df_XYcol}{Data, as 2 column data frame, where col.1 is X axis.}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The name of the file and title of the plot.}

\item{col}{Color of the plot.}

\item{palette_use}{GGpubr Color palette to use.}

\item{hide.legend}{hide legend}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{also.pdf}

\item{logX}{Make X axis log10-scale.}

\item{logY}{Make Y axis log10-scale.}

\item{xlab.angle}{Rotate X-axis labels by N degree.}

\item{hline}{Draw a horizontal line on the plot.}

\item{vline}{Draw a vertical line on the plot.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Scatter plot
}
\examples{
dfx <- as.data.frame(cbind("AA"=rnorm(500), "BB"=rnorm(500))); qscatter(dfx, suffix = "2D.gaussian")
}
