% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qstripchart}
\alias{qstripchart}
\title{qstripchart}
\usage{
qstripchart(
  df_XYcol_or_list,
  x = 1,
  y = 2,
  col = NULL,
  fill = "gold",
  plotname = FixPlotName(substitute(df_XYcol_or_list)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  ylab = NULL,
  plot = TRUE,
  add = c("violin", "mean_sd"),
  size.point = 0.2,
  stat.test = TRUE,
  stat.method = NULL,
  stat.label.y.npc = "top",
  stat.label.x = 0.75,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  hide.legend = FALSE,
  also.pdf = TRUE,
  save.obj = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  logY = FALSE,
  annotation_logticks_Y = logY,
  xlab.angle = 90,
  hline = FALSE,
  vline = FALSE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  grid = "y",
  max.categ = 100,
  w = qqqAxisLength(df_XYcol_or_list),
  h = 6,
  ...
)
}
\arguments{
\item{df_XYcol_or_list}{Data as a two-column data frame, where column 1 is the X axis, alternatively a uniquely named list of values.}

\item{x}{The index or name of the column to be plotted on the X axis. Default: \code{1}.}

\item{y}{The index or name of the column to be plotted on the Y axis. Default: \code{2}.}

\item{col}{The index or name of the column to be used for coloring the plot. Default: \code{NULL}.}

\item{fill}{Fill color of the plot. Default: \code{gold}.}

\item{plotname}{Name of the plot}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix.}

\item{filename}{Manually provided filename (optional). Default: parsed from \code{plotname}.}

\item{ylab}{Y-axis label. Default: NULL.}

\item{plot}{Display the plot.}

\item{add}{Add boxplot or violin chart? Default: add = c("violin", "mean_sd"); it can be "boxplot" or only "mean_sd".}

\item{size.point}{Size of points.}

\item{stat.test}{Do a statistical test?}

\item{stat.method}{stat method. NULL for default}

\item{stat.label.y.npc}{Stat label y position}

\item{stat.label.x}{Stat label x position}

\item{palette_use}{GGpubr color palette to use.}

\item{hide.legend}{Hide legend.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Save the ggplot object to a file. Default: FALSE.}

\item{ext}{File extension (.pdf / .png).}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{xlab.angle}{Rotate X-axis labels by N degrees. Default: 90}

\item{hline}{Draw a horizontal line on the plot.}

\item{vline}{Draw a vertical line on the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{max.categ}{The maximum allowed number of unique categories.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}
}
\description{
Generates a stripchart and saves the plot for a given 2-column dataframe and offers several customizations.
}
\examples{
data("ToothGrowth")
ToothLen.by.Dose <- ToothGrowth[, c("dose", "len")]
qstripchart(ToothLen.by.Dose)

}
