% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qstripchart}
\alias{qstripchart}
\title{qstripchart plot}
\usage{
qstripchart(
  df_XYcol_or_list,
  add = c("violin", "mean_sd"),
  suffix = NULL,
  plotname = FixPlotName(substitute(df_XYcol_or_list), suffix),
  title = F,
  size.point = 0.2,
  stat.test = T,
  stat.method = NULL,
  stat.label.y.npc = "top",
  stat.label.x = 0.75,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  hide.legend = FALSE,
  ext = "png",
  also.pdf = T,
  logY = FALSE,
  xlab.angle = 90,
  hline = FALSE,
  vline = FALSE,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = F),
  w = 7,
  h = w,
  ...
)
}
\arguments{
\item{df_XYcol_or_list}{Data, as 2 column data frame, where col.1 is X axis, alternatively a uniquely named list ov values.}

\item{add}{Add boxplot or violin chart? Default  add = c("violin", "mean_sd"), it can be "boxplot" or only "mean_sd".}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{Name of the plot}

\item{title}{The name of the file and title of the plot.}

\item{size.point}{Size of points}

\item{stat.test}{Do a statistical test?}

\item{stat.method}{stat method. NULL for default}

\item{stat.label.y.npc}{stat label y position}

\item{stat.label.x}{stat label x position}

\item{palette_use}{GGpubr Color palette to use.}

\item{hide.legend}{hide legend}

\item{ext}{File extension (.pdf / .png).}

\item{also.pdf}{also.pdf}

\item{logY}{Make Y axis log10-scale.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{hline}{Draw a horizontal line on the plot.}

\item{vline}{Draw a vertical line on the plot.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
qstripchart plot
}
\examples{
data("ToothGrowth"); ToothLen.by.Dose <- ToothGrowth[ ,c('dose', 'len')]; qstripchart(ToothLen.by.Dose)
}
