% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qvenn}
\alias{qvenn}
\title{qvenn - Venn Diagram}
\usage{
qvenn(
  list,
  also.pdf = FALSE,
  plotname = FixPlotName(substitute(list)),
  suffix = NULL,
  subtitle = paste(length(unique(unlist(list))), "elements in total"),
  caption = parseParamStringWNames(sapply(list, length)),
  caption2 = NULL,
  filename = NULL,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  col.min = "white",
  col.max = "red",
  hide.legend = FALSE,
  x_exp = 0.2,
  w = 8,
  h = 0.75 * w,
  ...
)
}
\arguments{
\item{list}{The variable to plot.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{subtitle}{The subtitle of the plot. Default: paste (length(unique(unlist(list))), 'elements in total')}

\item{caption}{Optional text added to bottom right corner of the plot. Default = List element lengths parsed.}

\item{caption2}{Optional text added to bottom right corner of the plot. Default = NULL.}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{col.min}{Color scale minimum, default: white}

\item{col.max}{Color scale maximum, default: red}

\item{hide.legend}{hide legend}

\item{x_exp}{Expand axis to show long set labels. Default: 0.2.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot. Default: h = w * 0.75.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Draw and save a Venn Diagram using the \code{ggVennDiagram} package.
}
\examples{
LetterSets <- list("One" = LETTERS[1:7], "Two" = LETTERS[3:12])
qvenn(LetterSets)

}
