% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qvenn}
\alias{qvenn}
\title{Venn Diagram}
\usage{
qvenn(
  list,
  ext = MarkdownHelpers::unless.specified("b.def.ext", def = "pdf"),
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  suffix = NULL,
  plotname = FixPlotName(kpp(substitute(list), suffix)),
  filename = FALSE,
  subtitle = paste(length(unique(unlist(list))), "elements in total"),
  col.min = "white",
  col.max = "red",
  hide.legend = FALSE,
  w = 8,
  h = 0.75 * w,
  ...
)
}
\arguments{
\item{list}{The variable to plot.}

\item{ext}{File extension (.pdf / .png).}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{suffix}{A suffix added to the filename. NULL by default.}

\item{plotname}{The title of the plot and the name of the file (unless specified in \code{filename}).}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{subtitle}{The subtitle of the plot. Default: paste (length(unique(unlist(list))), 'elements in total')}

\item{col.min}{Color scale minimum, default: white}

\item{col.max}{Color scale maximum, default: red}

\item{hide.legend}{hide legend}

\item{w}{width of the plot.}

\item{h}{height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Draw Venn Diagram using the  ggVennDiagram package.
}
\examples{
LetterSets <- list("One" = LETTERS[1:7], "Two" = LETTERS[3:12]); qvenn(LetterSets)
}
