% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpress.R
\name{qviolin}
\alias{qviolin}
\title{qviolin}
\usage{
qviolin(
  df_XYcol_or_list,
  plotname = FixPlotName(substitute(df_XYcol_or_list)),
  subtitle = NULL,
  suffix = NULL,
  caption = suffix,
  filename = NULL,
  stat.test = FALSE,
  stat.method = NULL,
  stat.label.y.npc = "top",
  stat.label.x = 0.5,
  palette_use = c("RdBu", "Dark2", "Set2", "jco", "npg", "aaas", "lancet", "ucscgb",
    "uchicago")[4],
  hide.legend = FALSE,
  also.pdf = FALSE,
  ext = MarkdownHelpers::ww.set.file.extension(default = "png", also_pdf = also.pdf),
  logY = FALSE,
  annotation_logticks_Y = logY,
  xlab.angle = 45,
  hline = FALSE,
  vline = FALSE,
  grid = FALSE,
  plot = TRUE,
  save = TRUE,
  mdlink = MarkdownHelpers::unless.specified("b.mdlink", def = FALSE),
  w = 7,
  h = w,
  ...
)
}
\arguments{
\item{df_XYcol_or_list}{Data, as 2 column data frame, where col.1 is X axis, alternatively a uniquely named list ov values.}

\item{plotname}{Name of the plot}

\item{subtitle}{Optional subtitle text added below the title. Default is NULL.}

\item{suffix}{Optional suffix added to the filename. Default is NULL.}

\item{caption}{Optional text added to bottom right corner of the plot. Default = suffix}

\item{filename}{Manually provided filename (optional). Default: parse from \code{plotname},}

\item{stat.test}{Do a statistical test?}

\item{stat.method}{stat method. NULL for default}

\item{stat.label.y.npc}{stat label y position}

\item{stat.label.x}{stat label x position}

\item{palette_use}{GGpubr Color palette to use.}

\item{hide.legend}{hide legend}

\item{also.pdf}{Save plot in both png and pdf formats}

\item{ext}{File extension (.pdf / .png).}

\item{logY}{Make Y axis log10-scale.}

\item{annotation_logticks_Y}{Logical indicating whether to add annotation logticks on Y-axis. Default follows the value of \code{logY}.}

\item{xlab.angle}{Rotate X-axis labels by N degree. Default: 90}

\item{hline}{Draw a horizontal line on the plot.}

\item{vline}{Draw a vertical line on the plot.}

\item{grid}{Character indicating the axis to add gridlines. Options are 'x', 'y', or 'xy'. Default is 'y'.}

\item{plot}{Display the plot.}

\item{save}{Save the plot into a file.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Pass any other parameter of the corresponding plotting function(most of them should work).}
}
\description{
Draw and save a violin plot
}
\examples{
data("ToothGrowth"); ToothLen.by.Dose <- ToothGrowth[ ,c('dose', 'len')]; qviolin(ToothLen.by.Dose)
}
