% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatter.R
\name{scatter_express}
\alias{scatter_express}
\title{Create a scatterplot and calculate correlation values for two numerical variables}
\usage{
scatter_express(df, xval = NA, yval = NA)
}
\arguments{
\item{df}{Dataframe to plot}

\item{xval}{x-var Column name used as the x-axis variable}

\item{yval}{y-var Column name used as the y-axis variable}
}
\value{
ggplot Object
}
\description{
Creates a ggplot scatterplot object containing two numerical variables. Arguements in the
function will control whether correlational values and log transformations are calculated for the input data.
}
