% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Time series decomposition and plot}
\usage{
ts_plot(data, col, type, frequency)
}
\arguments{
\item{data}{str: the path of the csv file}

\item{col}{str: the column to be analyzed}

\item{type}{str: "regular" or "irregular"
"regular": time series with regular time intervals
"irregular": time series with irregular time intervals}

\item{frequency}{int: the desired time interval}
}
\value{
a figure with 4 subplots
}
\description{
Convert csv file into a time series object, decompose it into in trend, seasonality/cyclicity
and noise/remainder/error and plot the raw data and the decomposed components.
}
\examples{
ts_plot("https://raw.githubusercontent.com/plotly/datasets/master/timeseries.csv", A, irregular, 3)
}
